Refactor searchable models
