+++ src/Models/Category.php
@@ -104,16 +104,11 @@ class Category extends Model
             ->get();
     }
 
+    /**
+     * @inheritdoc
+     */
     protected function makeAllSearchableUsing(Builder $query)
     {
-        // TODO: Is this filter still useful? You could override and
-        // extend the existing model which gives you full control.
-        // But... that will be applied always, this one is just
-        // for the index. Another option could be to have 2
-        // category models; a default and one specificly
-        // for the index: Models/IndexCategory.php
-        // directly handy to give these Scout
-        // methods their own place...
         return $query->withEventyGlobalScopes('index.categories.scopes')
             ->select((new (config('rapidez.models.category')))->qualifyColumns(['entity_id', 'name']))
             ->whereNotNull('url_key')
